refactor: remove WaitCompiler and associated helpers, streamlining internal compilation logic by eliminating unused WAITFOR expression handling
